ui: Move Statements reducers
Release note: None
