+++ pkg/ui/modules/statements/store/reducers/index.ts
@@ -1,0 +1,2 @@
+export * from './statementDiagnosticsReportsReducer'
+export * from './statementsReducer'
