+++ pkg/ui/modules/statements/views/index.ts
@@ -1,0 +1,2 @@
+export * from './statementsPage'
+export * from './diagnostics'
